+++ src/main/java/org/realityforge/gwt/eventsource/client/html5/Html5EventSource.java
@@ -3,6 +3,7 @@ package org.realityforge.gwt.eventsource.client.html5;
 import com.google.gwt.core.client.JavaScriptObject;
 import com.google.web.bindery.event.shared.EventBus;
 import com.google.web.bindery.event.shared.SimpleEventBus;
+import java.util.HashSet;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.realityforge.gwt.eventsource.client.EventSource;
@@ -14,7 +15,10 @@ public class Html5EventSource
     return !!window.EventSource;
   }-*/;
 
+  @Nullable
   private EventSourceImpl _eventSource;
+  @Nullable
+  private HashSet<String> _subscriptions;
 
   public static class Factory
     implements EventSource.Factory
@@ -34,14 +38,16 @@ public class Html5EventSource
   @Override
   public void close()
   {
-    checkConnected();
+    ensureEventSource();
     doClose();
   }
 
   private void doClose()
   {
+    assert null != _eventSource;
     _eventSource.close();
     _eventSource = null;
+    _subscriptions = null;
     onClose();
   }
 
@@ -53,6 +59,7 @@ public class Html5EventSource
       throw new IllegalStateException( "EventSource already connected" );
     }
     _eventSource = EventSourceImpl.create( this, url, withCredentials );
+    _subscriptions = new HashSet<String>();
   }
 
   @Nonnull
@@ -69,27 +76,68 @@ public class Html5EventSource
     }
   }
 
+  @Override
+  public boolean subscribeTo( @Nonnull final String messageType )
+    throws IllegalStateException
+  {
+    final HashSet<String> subscriptions = ensureSubscriptions();
+    if ( subscriptions.contains( messageType ) )
+    {
+      return false;
+    }
+    else
+    {
+      subscriptions.add( messageType );
+      ensureEventSource().subscribe( messageType );
+      return true;
+    }
+  }
+
+  @Override
+  public boolean unsubscribeFrom( @Nonnull final String messageType )
+    throws IllegalStateException
+  {
+    if ( ensureSubscriptions().remove( messageType ) )
+    {
+      ensureEventSource().unsubscribe( messageType );
+      return false;
+    }
+    else
+    {
+      return true;
+    }
+  }
+
   @Nonnull
   @Override
   public final String getURL()
   {
-    checkConnected();
-    return _eventSource.getURL();
+    return ensureEventSource().getURL();
   }
 
   @Override
   public final boolean getWithCredentials()
   {
-    checkConnected();
-    return _eventSource.getWithCredentials();
+    return ensureEventSource().getWithCredentials();
   }
 
-  private void checkConnected()
+  private EventSourceImpl ensureEventSource()
   {
     if ( null == _eventSource )
     {
       throw new IllegalStateException( "EventSource not open" );
     }
+    return _eventSource;
+  }
+
+  @Nonnull
+  protected final HashSet<String> ensureSubscriptions()
+  {
+    if ( null == _subscriptions )
+    {
+      throw new IllegalStateException( "EventSource not open" );
+    }
+    return _subscriptions;
   }
 
   private final static class EventSourceImpl
@@ -98,25 +146,27 @@ public class Html5EventSource
     static native EventSourceImpl create( EventSource client, String url, boolean withCredentials ) /*-{
       var eventSource = new EventSource( url, {withCredentials: withCredentials} );
       eventSource.onopen = $entry( function ()
-                          {
-                            client.@org.realityforge.gwt.eventsource.client.EventSource::onOpen()();
-                          } );
+                                   {
+                                     client.@org.realityforge.gwt.eventsource.client.EventSource::onOpen()();
+                                   } );
       eventSource.onerror = $entry( function ()
-                           {
-                             if ( eventSource.readyState == EventSource.CLOSED )
-                             {
-                               // Connection was closed.
-                               client.@org.realityforge.gwt.eventsource.client.html5.Html5EventSource::doClose()();
-                             }
-                             else
-                             {
-                               client.@org.realityforge.gwt.eventsource.client.EventSource::onError()();
-                             }
-                           } );
+                                    {
+                                      if ( eventSource.readyState == EventSource.CLOSED )
+                                      {
+                                        // Connection was closed.
+                                        client.@org.realityforge.gwt.eventsource.client.html5.Html5EventSource::doClose()();
+                                      }
+                                      else
+                                      {
+                                        client.@org.realityforge.gwt.eventsource.client.EventSource::onError()();
+                                      }
+                                    } );
       eventSource.onmessage = $entry( function ( response )
-                             {
-                               client.@org.realityforge.gwt.eventsource.client.EventSource::onMessage(Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)( response.lastEventId, response.type, response.data );
-                             } );
+                                      {
+                                        client.@org.realityforge.gwt.eventsource.client.EventSource::onMessage(Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)( response.lastEventId,
+                                                                                                                                                                        response.type,
+                                                                                                                                                                        response.data );
+                                      } );
       return eventSource;
     }-*/;
 
@@ -138,6 +188,14 @@ public class Html5EventSource
 
     native void close() /*-{
       this.close();
+    }-*/;
+
+    native void subscribe( String messageType ) /*-{
+      this.addEventListener( messageType, this.onmessage );
+    }-*/;
+
+    native void unsubscribe( String messageType ) /*-{
+      this.removeEventListener( messageType, this.onmessage );
     }-*/;
   }
 }