+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.4:
 
+* Support a variant of EventSource.open() that sets withCredentials option.
+* Rename EventSource.connect() to EventSource.open() to match terminology of underlying construct.
 * Add support for EventSource.getURL() and EventSource.getWithCredentials() methods.
 
 ## 0.3: